sessions

+++ pages/courses/courses.php
@@ -35,9 +35,11 @@
                                 echo '<td width=250>';
                                 echo '<a class="btn" href="courses_read.php?course_id='.$row['course_id'].'">Read</a>';
                                 echo ' ';
+                                if($_SESSION['is_admin']=1) {
                                 echo '<a class="btn btn-success" href="courses_edit.php?course_id='.$row['course_id'].'">Промени</a>';
                                 echo ' ';
                                 echo '<a class="btn btn-danger" href="courses_delete.php?course_id='.$row['course_id'].'">Изтрий</a>';
+								}
                                 echo '</td>';
                                 echo '</tr>';
                        }
